sort in conversation list component
